requirements.txt
# Speech-to-Text with OpenAI Whisper - Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
openai-whisper>=20231117
sounddevice>=0.4.6
numpy>=1.21.0

# Audio file support
librosa>=0.10.0
soundfile>=0.12.1

# Optional: For better audio processing
scipy>=1.9.0

